+++ src/commonTest/kotlin/kotools/types/experimental/NegativeIntTest.kt
@@ -7,9 +7,11 @@ package kotools.types.experimental
 
 import kotools.types.number.NegativeInt
 import kotools.types.number.PositiveInt
+import kotools.types.number.toNegativeInt
 import kotools.types.number.toPositiveIntOrFailure
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
 class NegativeIntTest {
     @ExperimentalKotoolsTypesApi
@@ -21,5 +23,18 @@ class NegativeIntTest {
             .unaryMinus()
             .toPositiveIntOrFailure()
         assertEquals(expected, actual)
+    }
+
+    @ExperimentalKotoolsTypesApi
+    @Test
+    fun range_should_go_from_Int_MIN_VALUE_included_to_zero_included() {
+        val actual: NotEmptyRange<NegativeInt> = NegativeInt.range
+        assertTrue { actual.start is InclusiveBound }
+        val expectedStartValue: NegativeInt = Int.MIN_VALUE.toNegativeInt()
+            .getOrThrow()
+        assertEquals(expectedStartValue, actual.start.value)
+        val expectedEndValue: NegativeInt = 0.toNegativeInt()
+            .getOrThrow()
+        assertEquals(expectedEndValue, actual.end.value)
     }
 }